requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.19.0
packaging==22.0
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.7.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
